chore(tktrex): refactored downloader parser into typescript, and restored its operation
